+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/_client_factory.py
@@ -14,13 +14,17 @@ def usage_details_mgmt_client_factory(cli_ctx, kwargs):
     return cf_consumption(cli_ctx, **kwargs).usage_details
 
 
-def reservations_summaries_mgmt_client_factory(cli_ctx, kwargs):
+def reservation_summary_mgmt_client_factory(cli_ctx, kwargs):
     return cf_consumption(cli_ctx, **kwargs).reservations_summaries
 
 
-def reservations_details_mgmt_client_factory(cli_ctx, kwargs):
+def reservation_detail_mgmt_client_factory(cli_ctx, kwargs):
     return cf_consumption(cli_ctx, **kwargs).reservations_details
 
 
 def pricesheet_mgmt_client_factory(cli_ctx, kwargs):
     return cf_consumption(cli_ctx, **kwargs).price_sheet
+
+
+def marketplace_mgmt_client_factory(cli_ctx, kwargs):
+    return cf_consumption(cli_ctx, **kwargs).marketplaces